requirements.txt
streamlit
google-generativeai
SpeechRecognition
moviepy
numpy